fix Track - delete 'm' in filed names fix RealmRepository - bad id added save results in realm and show them from Realm in ResultFragment

+++ app/src/main/java/com/elegion/tracktor/utils/StringUtils.java
@@ -10,7 +10,7 @@ public class StringUtils {
         return String.format(Locale.ENGLISH, "%.1f m", distance);
     }
 
-    public static String getTimerText(int totalSeconds) {
+    public static String getTimerText(long totalSeconds) {
         long hours = totalSeconds / 3600;
         long minutes = (totalSeconds % 3600) / 60;
         long seconds = totalSeconds % 60;
